+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- [#694](https://github.com/spegel-org/spegel/pull/694) Replace IP in multi address with manet.
+
 ### Deprecated
 
 ### Removed